Lots of visual updates, some js and css updates.
decided on a new font, added functionality to button that causes it to
be 25% opacity when a cookie is not able to be made from ingredients in
pot, but trying to improve, commented out working version and have 'fix'
live, doesn't fully work yet. Still seems to work as well as prior
version, though. Font switcher button only applies to H2's, and
functions. Cleaned up some spare CSS/html stuff.