NEW: attempt to create dual interface
Added a 'wrapper' to hold two similar divs, and the beginning of an
interface to display the algorithm. Also adjusted the font switcher
button position a bit, some CSS changes and experiments, and adding the
algorithm file itself of course, along with some edits to prepare it for
jQuery use.