FIXED BUTTON. Entirely rebuilt.
And other small things.

Weird bug... float:left was producing it. Lots of things actually can
produce it, related to positioning, for some reason. position:absolute
and some top/left px numbers produced it, as well as injecting it after
some divs (but not all). Set it to show on bottom of page unobtrusively
for now.

Also, prettier hr's.

Was a nightmare to find this bug, since it didn't make sense. Entirely
rebuilt the button to inject itself from scratch on page load. Still had
the problem. I then put the code on a codeacademy page, and noticed it
all worked there.

So now I knew the problem wasn't in the button code.

Because it was all now contained in a single file, and I realized the
problem had to be elsewhere, I was able to test: html, other js doc, or
css doc? by removing almost all html and both other js docs and css doc,
verifying it worked like that, and then slowly adding things back in.

That got me to it being in the CSS. So then I opened up the browser
console and went through and manually adjusted css elements until I
found the one (float:left) that unchecking fixed.

Voila.