cleaned up algchal, + enter key feature (w/glitch...)
Erased old prompt() functionality, cleaned up algchal doc. Added
detection for enter key, separated out the functionality that button
activated into a new function, and made either a button click or the
enter key run that function.

Problem is, there's a glitch (in Chrome for sure, not sure others yet):
it runs as expected, but then it clears itself before anything is
visible. The only way I can know this for sure is that if I set an alert
box to run at the beginning of the function, closing the alert box
triggers a lag of a fraction of a second to make it clear that it is
running before it clears itself.

Weird.